requirement.txt
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.11.5
click==6.7
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pycparser==2.18
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
WTForms==2.1
google-api-core==1.14.3
google-api-python-client==1.8.3
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.1.0
google-cloud-language==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.23.0
google-cloud-texttospeech==1.0.1
google-cloud-translate==2.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
pydub
wave
mutagenpip3 
